save large code refactoring

+++ src/main/java/com/rekahdo/ecommerce/goodshop/_repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.rekahdo.ecommerce.goodshop._repository;
 import com.rekahdo.ecommerce.goodshop._entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Byte> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
